move the delete function into Book.js from List.js
