+++ modules/BookController.js
@@ -18,17 +18,4 @@ export default class BooktoDom {
           </tr>          
           `;
   }
-
-  static checkStorage(target) {
-    const dataStored = localStorage.getItem('books');
-    if (dataStored) {
-      const books = JSON.parse(dataStored);
-      target.innerHTML = '';
-      for (let key in books) {
-        target.innerHTML += BooktoDom.append(key, books[key]);
-      }
-    } else {
-      target.innerHTML = `<tr> <td colspan="4">Nothing to show </td> <tr>`;
-    }
-  }
 }
